Delete confirmation added

+++ src/app/posts/posts-list/posts-list.component.ts
@@ -23,9 +23,11 @@ export class PostsListComponent implements OnInit {
   }
 
   deletePost(thePost) {
-    console.log(thePost);
-    this._store.dispatch(
-      deleteExistingPostAction({ existingPostToBeDeleted: thePost })
-    );
+    if (confirm('Are you sure want to delete the post?')) {
+      console.log(thePost);
+      this._store.dispatch(
+        deleteExistingPostAction({ existingPostToBeDeleted: thePost })
+      );
+    }
   }
 }
